+++ bindings/flutter/packages/runanywhere/ios/runanywhere/Sources/runanywhere/RunAnywherePlugin.swift
@@ -39,9 +39,9 @@ public class RunAnywherePlugin: NSObject, FlutterPlugin {
         case "getPlatformVersion":
             result("iOS " + UIDevice.current.systemVersion)
         case "getSDKVersion":
-            result("0.20.27")
+            result("0.20.28")
         case "getCommonsVersion":
-            result("0.20.27")
+            result("0.20.28")
         default:
             result(FlutterMethodNotImplemented)
         }